fix not close bug
